`Refactor BaseGenerator to use frontendGeneratorConfig instead of config`
