+++ Sistema.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Sistema = () => {
+	return <div>Sistema component generated successfully!</div>;
+};
+
+export default Sistema;
