+++ src/base-generator.ts
@@ -27,4 +27,15 @@ export abstract class BaseGenerator {
             console.error(`Error writing file at ${filePath}: ${error}`);
         }
     }
+
+    protected loadMetadata(): any {
+        try {
+            const metadataPath = path.join(this.frontendGeneratorConfig.outputDir, 'request-response-metadata.json');
+            const metadata = fs.readFileSync(metadataPath, 'utf-8');
+            return JSON.parse(metadata);
+        } catch (error) {
+            console.error("Failed to load metadata:", error);
+            return null;
+        }
+    }
 }
