+++ src/component-generator.ts
@@ -33,9 +33,6 @@ const ${this.frontendConfig.app} = () => {
 export default ${this.frontendConfig.app};
 `;
 
-        const filePath = path.join(outputDir, `${this.frontendConfig.app}.tsx`);
-        fs.writeFileSync(filePath, componentContent.trim());
-
-        console.log(`Component generated at ${filePath}`);
+        fs.writeFileSync(`${this.frontendConfig.app}.tsx`, componentContent.trim());
     }
 }
