+++ src/context-generator.ts
@@ -1,7 +1,5 @@
 // src/context-generator.ts
 
-import * as fs from 'fs';
-import * as path from 'path';
 import { IGenerator, FrontendGeneratorConfig, RequestConfig } from './interfaces';
 import { BaseGenerator } from './base-generator';
